Check UID & GID of "domjudge" are same across Docker build stages
Since commit b6ae2719d4 (Preserve ownership of /opt/domjudge set by
`make install-*`, 2024-07-10), the ownership of the files copied from
the previous build stage is preserved rather than overwritten. This is
slightly fragile because Docker's `COPY --from` directive only
preserves the ownership correctly if the numerical user and group IDs
are the same across the build stages.

This commit adds a check to verify that the IDs of the "domjudge" user
are the same. (We don't need to check the IDs of the "www-data" user,
because they're guaranteed to be the same.[1][2])

 - For the domserver image, we do this by comparing the values in
   /etc/passwd.

 - For the judgehost image, we can't compare /etc/passwd because the
   first build stage is in a separate Dockerfile (chroot-and-tar.sh
   needs to run in a privileged container). So instead we check the
   owner of a well-known directory (/opt/domjudge/judgehost/run) which
   should be owned by "domjudge". (We can't check the GID because
   there are no files or directories with group "domjudge".)

This commit also moves the creation of the "domjudge" user a little
earlier, to reduce the opportunity for some other command to create a
user and change the ID assigned to the "domjudge" user.

(An alternative approach that guarantees the ownership is preserved,
regardless of the UID & GID, is to create a tarball, copy the tarball
between the build stages, then extract the tarball. This is reliable
because GNU tar records the user name and attempts to preserve it when
extracting. However it's slower and makes the image larger.)

[1] https://www.debian.org/doc/debian-policy/ch-opersys.html#uid-and-gid-classes
[2] https://salsa.debian.org/debian/base-passwd/blob/master/passwd.master